refactor cloudProviders for provisioning
