+++ validation/provisioning/rke2/hardened_cluster_test.go
@@ -34,7 +34,6 @@ type HardenedRKE2ClusterProvisioningTestSuite struct {
 	provisioningConfig  *provisioninginput.Config
 	project             *management.Project
 	chartInstallOptions *charts.InstallOptions
-	chartFeatureOptions *charts.RancherMonitoringOpts
 }
 
 func (c *HardenedRKE2ClusterProvisioningTestSuite) TearDownSuite() {
