+++ validation/upgrade/cloud_provider_version.go
@@ -8,7 +8,7 @@ import (
 	"github.com/rancher/shepherd/extensions/clusters"
 	"github.com/rancher/shepherd/extensions/workloads/pods"
 	"github.com/rancher/tests/actions/charts"
-	"github.com/rancher/tests/actions/provisioning/permutations"
+	"github.com/rancher/tests/actions/cloudprovider"
 	"github.com/rancher/tests/actions/provisioninginput"
 	"github.com/sirupsen/logrus"
 	"github.com/stretchr/testify/require"
@@ -31,5 +31,5 @@ func upgradeVsphereCloudProviderCharts(t *testing.T, client *rancher.Client, clu
 	clusterObject, err := adminClient.Steve.SteveType(clusters.ProvisioningSteveResourceType).ByID(provisioninginput.Namespace + "/" + clusterID)
 	require.NoError(t, err)
 
-	permutations.CreatePVCWorkload(t, client, clusterObject)
+	cloudprovider.CreatePVCWorkload(t, client, clusterObject)
 }
