+++ validation/upgrade/cloud_provider_version_test.go
@@ -14,9 +14,8 @@ import (
 
 type UpgradeCloudProviderSuite struct {
 	suite.Suite
-	session  *session.Session
-	client   *rancher.Client
-	clusters []string
+	session *session.Session
+	client  *rancher.Client
 }
 
 func (u *UpgradeCloudProviderSuite) TearDownSuite() {
